+++ src/Lab4/Commands/FileDeleteCommand.cs
@@ -1,3 +1,4 @@
+using System;
 using System.IO;
 
 namespace Itmo.ObjectOrientedProgramming.Lab4.Commands;
@@ -13,6 +14,12 @@ public class FileDeleteCommand : ICommand
 
     public void Execute()
     {
+        if (!File.Exists(_path))
+        {
+            Console.WriteLine("This path doesn't exist");
+            return;
+        }
+
         File.Delete(_path);
     }
 }
